Run script to install Alacritty config file
Creating a symbolic link from WSL to Windows doesn't work for this
purpose. While the symbolic link is created, it's only detected as such
within WSL. Explorer sees it as a plain, empty file. Upon launching
Alacritty, it states that "the file cannot be accessed by the system".

Using a script that simply copies the config file to the correct
location is a good alternative. The only downside is dotbot's install
script needs to be executed every time the config file is modified.